# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Survivor Sharks API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pools:
    get:
      operationId: listPools
      tags: [pools]
      summary: List pools the current user belongs to
      responses:
        "200":
          description: List of pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pool"

    post:
      operationId: createPool
      tags: [pools]
      summary: Create a new survivor pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoolInput"
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
  /pools/join:
    post:
      operationId: joinPool
      tags: [pools]
      summary: Join a pool using an invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinPoolInput"
      responses:
        "200":
          description: Joined pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
        "404":
          description: Pool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pools/{poolId}:
    get:
      operationId: getPool
      tags: [pools]
      summary: Get pool details
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolDetail"
        "404":
          description: Pool not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePool
      tags: [pools]
      summary: Update pool settings (commissioner only)
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoolUpdate"
      responses:
        "200":
          description: Updated pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"

    delete:
      operationId: deletePool
      tags: [pools]
      summary: Delete a pool (commissioner or admin only)
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Picks ─────────────────────────────────────────────────────────────────
...
  /pools/{poolId}/results:
    get:
      operationId: getResults
      tags: [results]
      summary: Get all weekly results for a pool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeekResult"

    post:
      operationId: processResults
      tags: [results]
      summary: Process weekly results and eliminate losers (commissioner only)
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessResultsInput"
      responses:
        "200":
          description: Results processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResultsResult"

  # ── Eliminations / Kill History ───────────────────────────────────────────
...
  /pools/{poolId}/stats:
    get:
      operationId: getPoolStats
      tags: [pools]
      summary: Get pool statistics summary
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolStats"

  # ── Sports / Teams / Schedule ─────────────────────────────────────────────
...
  /admin/pools/{poolId}:
    delete:
      operationId: adminDeletePool
      tags: [admin]
      summary: Admin — force delete any pool
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth Schemas ──────────────────────────────────────────────────────────
...
    PoolInput:
      type: object
      required: [name, sport]
      properties:
        name:
          type: string
        sport:
          type: string
          enum: [nfl, mlb, nba, nhl, fifa]
        description:
          type: string
        maxEntries:
          type: integer
        entryFee:
          type: number
        prizePot:
          type: number
        currentWeek:
          type: integer
        season:
          type: integer
    PoolUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        maxEntries:
          type: integer
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
    JoinPoolInput:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string
    Pool:
      type: object
      required: [id, name, sport, inviteCode, currentWeek, season, isActive, memberCount, commissionerId]
      properties:
        id:
          type: integer
        name:
          type: string
        sport:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
        memberCount:
          type: integer
        activeCount:
          type: integer
        commissionerId:
          type: integer
        commissionerName:
          type: string
        maxEntries:
          type: ["integer", "null"]
        entryFee:
          type: ["number", "null"]
        prizePot:
          type: ["number", "null"]
        createdAt:
          type: string
    PoolDetail:
      type: object
      required: [id, name, sport, inviteCode, currentWeek, season, isActive, members, commissionerId]
      properties:
        id:
          type: integer
        name:
          type: string
        sport:
          type: string
        description:
          type: ["string", "null"]
        inviteCode:
          type: string
        currentWeek:
          type: integer
        season:
          type: integer
        isActive:
          type: boolean
        commissionerId:
          type: integer
        commissionerName:
          type: string
        maxEntries:
          type: ["integer", "null"]
        entryFee:
          type: ["number", "null"]
        prizePot:
          type: ["number", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/PoolMember"
        createdAt:
          type: string
    PoolMember:
      type: object
      required: [userId, username, status, joinedAt]
      properties:
        userId:
          type: integer
        username:
          type: string
        displayName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, eliminated]
        eliminatedWeek:
          type: ["integer", "null"]
        joinedAt:
          type: string
    PoolStats:
      type: object
      required: [poolId, totalMembers, activeMembers, eliminatedMembers, currentWeek, weeklyPickRate, mostPickedTeam]
      properties:
        poolId:
          type: integer
        totalMembers:
          type: integer
        activeMembers:
          type: integer
        eliminatedMembers:
          type: integer
        currentWeek:
          type: integer
        weeklyPickRate:
          type: number
        mostPickedTeam:
          type: ["string", "null"]
        mostPickedTeamCount:
          type: ["integer", "null"]
        survivorPercentage:
          type: number

    # ── Pick Schemas ──────────────────────────────────────────────────────────
...
    ProcessResultsInput:
      type: object
      required: [week, losingTeamIds]
      properties:
        week:
          type: integer
        losingTeamIds:
          type: array
          items:
            type: string
    ProcessResultsResult:
      type: object
      required: [week, eliminated, survived, processedAt]
      properties:
        week:
          type: integer
        eliminated:
          type: array
          items:
            type: string
        survived:
          type: array
          items:
            type: string
        processedAt:
          type: string
    WeekResult:
      type: object
      required: [id, poolId, week, losingTeamIds, processedAt]
      properties:
        id:
          type: integer
        poolId:
          type: integer
        week:
          type: integer
        losingTeamIds:
          type: array
          items:
            type: string
        processedAt:
          type: string
        processedBy:
          type: integer
...